+++ dimensional/src/lib.ts
@@ -1,8 +1,31 @@
+import { Dimension } from './dimension';
+import { Quantity } from './quantity';
+import { Unit } from './unit';
+
 /**
- * Represents a type where every object value is a string.
+ * Shorthand for creating a set of physical base dimensions.
+ * @param exponents Exponents of each of the dimensions
+ * @returns A new dimension object
  */
-export type Dictionary<T extends string> = { [index in T]: string };
+export function D(exponents: Dimension.Exponents): Dimension.Dimension {
+    return new Dimension.Dimension(exponents);
+}
+
 /**
- * Represents a type where every object value is a number.
+ * Shorthand for creating a new unit of measurement.
+ * @param exponents Exponents on each of the individual units
+ * @returns A new unit object
  */
-export type NumberDictionary<T extends string> = { [index in T]?: number };
+export function U(exponents: Unit.Exponents): Unit.Unit {
+    return new Unit.Unit(exponents);
+}
+
+/**
+ * Shorthand for creating a new measured quantity.
+ * @param value The numerical value of the measurement
+ * @param unit The physical units of the measurement
+ * @returns A new quantity object
+ */
+export function Q(value: number, unit: Unit.Unit): Quantity.Quantity {
+    return new Quantity.Quantity(value, unit);
+}