+++ dimensional/src/unit.ts
@@ -1,41 +1,63 @@
 import { Compound } from './compound';
-import { ConversionTable } from './conversion';
+import { Conversion } from './conversion';
 import { Dimension } from './dimension';
-import { NumberDictionary } from './lib';
-
 /**
- * Contains a list of all units related to time.
+ * Contains all software used for the calculation of units of measurement.
  */
-export type TimeUnits = 'nanoseconds' | 'milliseconds' | 'seconds' | 'minutes' | 'hours' | 'days' | 'weeks' | 'months' | 'years';
-/**
- * Contains a list of all units related to distance.
- */
-export type LengthUnits = 'nanometers' | 'micrometers' | 'microns' | 'millimeters' | 'centimeters' | 'meters' | 'kilometers' | 'inches' | 'feet' | 'yards' | 'miles';
-/**
- * Contains a list of all unit sub-lists.
- */
-export type Units = TimeUnits | LengthUnits;
-/**
- * Defines the class for units for physical quantities.
- */
-export class Unit extends Compound<Units, Unit> {
-    public readonly dimension: Dimension;
-    constructor(exponents: NumberDictionary<Units>) {
-        super(exponents, t => ConversionTable[t].latex);
-        this.dimension = new Dimension({});
-        for (const unit of this.getNonzeroExponents()) {
-            this.dimension = this.dimension.mult(ConversionTable[unit].dim, this.getExponent(unit));
+export namespace Unit {
+    /**
+     * Contains a list of all physical units of measurement.
+     */
+    export type Name =
+        // Time span/duration
+        | 'nanosecond' | 'millisecond' | 'second' | 'minute' | 'hour' | 'day' | 'week' | 'month' | 'year'
+        // Length/distance
+        | 'nanometer' | 'micrometer' | 'micron' | 'millimeter' | 'centimeter' | 'meter' | 'kilometer' | 'inch' | 'foot' | 'yard' | 'mile'
+        // Mass
+        | 'milligram' | 'gram' | 'kilogram' | 'tonne' | 'ounce' | 'pound_mass' | 'slug' | 'stone' | 'shortton'
+        // Current
+        | 'milliampere' | 'ampere'
+        // Temperature
+        | 'Kelvin' | 'Rankine' | 'Celsius_rel' | 'Fahrenheight_rel'
+        // Substance
+        | 'mole'
+        // Luminous Intensity
+        | 'candela' | 'lumen'
+        // Force
+        | 'Newton' | 'kiloNewton' | 'pound_force'
+        // Trailing semicolon
+        ;
+    /**
+     * Is an object containing keys of units and values of nonzero exponents.
+     */
+    export interface Exponents extends Compound.Exponents<Name> { };
+    /**
+     * Defines the class for units of measurement for physical quantities.
+     */
+    export class Unit extends Compound.Compound<Name, Unit> {
+        public readonly dimension: Dimension.Dimension;
+        public readonly scale: number;
+        constructor(exponents: Exponents) {
+            const conversion: Conversion.Conversion = Unit.getConversion(exponents, 1, Dimension.None);
+            super(exponents, t => Conversion.Table[t]().latex);
+            this.scale = conversion.scale;
+            this.dimension = conversion.dimension;
+        }
+        public mult(other: Unit, exponent: number): Unit {
+            return new Unit(this.combine(other, exponent));
+        }
+        private static getConversion(exponents: Exponents, scale: number, dimension: Dimension.Dimension): Conversion.Conversion {
+            for (const unit in exponents) {
+                const conversion: Conversion.Conversion = Conversion.Table[unit as Unit.Name](),
+                    exponent: number = exponents[unit as Unit.Name] ?? 0;
+                scale *= (conversion.scale ** exponent);
+                dimension = dimension.mult(conversion.dimension, exponent);
+            }
+            return new Conversion.Conversion('', scale, dimension);
         }
     }
-    public mult(other: Unit, exponent: number): Unit {
-        return new Unit(this.combine(other, exponent));
-    }
-}
-/**
- * Shorthand for creating a new unit object.
- * @param exponents Exponents on each of the individual units
- * @returns A new unit object
- */
-export function U(exponents: NumberDictionary<Units>): Unit {
-    return new Unit(exponents);
+    /**
+     * Represents a unitless value.
+     */
+    export const None = new Unit({});
 }